stats/requirements.txt
aiokafka==0.10.0
clickhouse-connect==0.8.17
fastapi==0.115.0
grpcio==1.71.0
grpcio-tools
asyncpg
fastapi
uvicorn
asyncio
python-dateutil>=2.8.2